+++ depNotify.sh
@@ -1,24 +1,20 @@
 #!/bin/bash
-
-#########################################################################################
-# Version 1.2.0
-#########################################################################################
+# Version 2.0.0
 
 #########################################################################################
 # License information
 #########################################################################################
-
 # Copyright 2018 Jamf Professional Services
-#
+
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this
 # software and associated documentation files (the "Software"), to deal in the Software
 # without restriction, including without limitation the rights to use, copy, modify, merge,
 # publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
 # to whom the Software is furnished to do so, subject to the following conditions:
-#
+
 # The above copyright notice and this permission notice shall be included in all copies or
 # substantial portions of the Software.
-#
+
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 # INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
 # PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
@@ -29,7 +25,6 @@
 #########################################################################################
 # General Information
 #########################################################################################
-
 # This script is designed to make implementation of DEPNotify very easy with limited
 # scripting knowledge. The section below has variables that may be modified to customize
 # the end user experience. DO NOT modify things in or below the CORE LOGIC area unless
@@ -38,39 +33,24 @@
 # More information at: https://github.com/jamfprofessionalservices/DEP-Notify
 
 #########################################################################################
-# Visual Appearance and General Functionality Variables to Modify
+# Testing Mode
 #########################################################################################
-
 # Testing flag will enable the following things to change:
 # Auto removal of BOM files to reduce errors
 # Sleep commands instead of policies or other changes being called
 # Quit Key set to command + control + x
   TESTING_MODE=true # Set variable to true or false
 
+#########################################################################################
+# General Appearance
+#########################################################################################
 # Flag the app to open fullscreen or as a window
   FULLSCREEN=true # Set variable to true or false
 
-# Flag script to keep the computer from sleeping. BE VERY CAREFUL WITH THIS FLAG!
-# This flag could expose your data to risk by leaving an unlocked computer wide open.
-# Only recommended if you are using fullscreen mode and have a logout taking place at
-# the end of configuration (like for FileVault). Some folks may use this in workflows
-# where IT staff are the primary people setting up the device. The device will be
-# allowed to sleep again once the DEPNotify app is quit as caffeinate is looking
-# at DEPNotify's process ID.
-  NO_SLEEP=false
-
 # Banner image can be 600px wide by 100px high. Images will be scaled to fit
-# If this variable is left blank, the generic image will appear
+# If this variable is left blank, the generic image will appear. If using custom Self
+# Service branding, please see the Customized Self Service Branding area below
   BANNER_IMAGE_PATH="/Applications/Self Service.app/Contents/Resources/AppIcon.icns"
-
-  # Flag for using the custom branding icon from Self Service and Jamf Pro
-  # This will override the banner image specified above. If you have changed the
-  # name of Self Service, make sure to modify the Self Service name below.
-    SELF_SERVICE_CUSTOM_BRANDING=false # Set variable to true or false
-
-    # If using a name other than Self Service with Custom branding. Change the
-    # name with the SELF_SERVICE_APP_NAME variable below. Keep .app on the end
-      SELF_SERVICE_APP_NAME="Self Service.app"
 
 # Main heading that will be displayed under the image
 # If this variable is left blank, the generic banner will appear
@@ -80,11 +60,6 @@
 # If this variable is left blank, the generic message will appear. Leave single
 # quotes below as double quotes will break the new lines.
   MAIN_TEXT='Thanks for choosing a Mac at Organization! We want you to have a few applications and settings configured before you get started with your new Mac. This process should take 10 to 20 minutes to complete. \n \n If you need additional software or help, please visit the Self Service app in your Applications folder or on your Dock.'
-
-# URL for support or help that will open when the ? is clicked
-# If this variable is left blank, the ? will not appear
-# If using fullscreen mode, Safari will be launched behind the DEP Notify window
-  SUPPORT_URL="https://support.apple.com"
 
 # Initial Start Status text that shows as things are firing up
   INITAL_START_STATUS="Initial Configuration Starting..."
@@ -110,15 +85,32 @@
     COMPLETE_ALERT_TEXT="Your Mac is now finished with initial setup and configuration. Press Quit to get started!"
   # Options if not using dropdown alert box
     COMPLETE_MAIN_TEXT='Your Mac is now finished with initial setup and configuration.'
-    COMPLETE_BUTTON_TEXT="Get Started"
+    COMPLETE_BUTTON_TEXT="Get Started!"
 
-# If using EULA or Registration Window below, this will configure where the file is saved.
-# You may want to save the file for purposes like verifying EULA acceptance. This variable
-# is in a function so that $CURRENT_USER can be used. Variable must end with a slash.
-  INFO_PLIST_WRAPPER (){
-    DEP_NOTIFY_INFO_PLIST_PATH="/Users/$CURRENT_USER/Library/Preferences/"
-  }
+#########################################################################################
+# Plist Configuration
+#########################################################################################
+# The menu.depnotify.plist contains more and more things that configure the DEPNotify app
+# You may want to save the file for purposes like verifying EULA acceptance or validating
+# other options.
 
+# Plist Save Location
+  # This wrapper allows variables that are created later to be used but also allow for
+  # configuration of where the plist is stored
+    INFO_PLIST_WRAPPER (){
+      DEP_NOTIFY_USER_INPUT_PLIST="/Users/$CURRENT_USER/Library/Preferences/menu.nomad.DEPNotifyUserInput.plist"
+    }
+
+# Status Text Alignment
+  # The status text under the progress bar can be configured to be left, right, or center
+    STATUS_TEXT_ALIGN="center"
+
+# Help Button Configuration
+  # The help button was changed to a popup. Button will appear if title is populated.
+    HELP_BUBBLE_TITLE="Need Help?"
+    HELP_BUBBLE_BODY="This tool at Organization is designed to help with new employee onboarding. If you have issues, please give us a call at 123-456-7890"
+
+#########################################################################################
 # Error Screen Text
 #########################################################################################
 # If testing mode if false and configuration files are present, this text will appear to
@@ -141,7 +133,6 @@
 #########################################################################################
 # Policy Variable to Modify
 #########################################################################################
-
 # The policy array must be formatted "Progress Bar text,customTrigger". These will be
 # run in order as they appear below.
   POLICY_ARRAY=(
@@ -158,11 +149,49 @@
   )
 
 #########################################################################################
+# Caffeinate / No Sleep Configuration
+#########################################################################################
+# Flag script to keep the computer from sleeping. BE VERY CAREFUL WITH THIS FLAG!
+# This flag could expose your data to risk by leaving an unlocked computer wide open.
+# Only recommended if you are using fullscreen mode and have a logout taking place at
+# the end of configuration (like for FileVault). Some folks may use this in workflows
+# where IT staff are the primary people setting up the device. The device will be
+# allowed to sleep again once the DEPNotify app is quit as caffeinate is looking
+# at DEPNotify's process ID.
+  NO_SLEEP=false
+
+#########################################################################################
+# Customized Self Service Branding
+#########################################################################################
+# Flag for using the custom branding icon from Self Service and Jamf Pro
+# This will override the banner image specified above. If you have changed the
+# name of Self Service, make sure to modify the Self Service name below.
+# Please note, custom branding is downloaded from Jamf Pro after Self Service has opened
+# at least one time. The script is designed to wait until the files have been downloaded.
+# This could take a few minutes depending on server and network resources.
+  SELF_SERVICE_CUSTOM_BRANDING=false # Set variable to true or false
+
+# If using a name other than Self Service with Custom branding. Change the
+# name with the SELF_SERVICE_APP_NAME variable below. Keep .app on the end
+  SELF_SERVICE_APP_NAME="Self Service.app"
+
+#########################################################################################
 # EULA Variables to Modify
 #########################################################################################
-
 # EULA configuration
   EULA_ENABLED=false # Set variable to true or false
+
+  # EULA status bar text
+    EULA_STATUS="Waiting on completion of EULA acceptance"
+
+  # EULA button text on the main screen
+    EULA_BUTTON="Read and Agree to EULA"
+
+  # EULA Screen Title
+    EULA_MAIN_TITLE="Organization End User License Agreement"
+
+  # EULA Subtitle
+    EULA_SUBTITLE="Please agree to the following terms and conditions to start configuration of this Mac"
 
   # Path to the EULA file you would like the user to read and agree to. It is
   # best to package this up with Composer or another tool and deliver it to a
@@ -172,112 +201,219 @@
 #########################################################################################
 # Registration Variables to Modify
 #########################################################################################
-
 # Registration window configuration
-  REGISTER_ENABLED=false # Set variable to true or false
+  REGISTRATION_ENABLED=false # Set variable to true or false
 
   # Registration window title
-    REGISTER_TITLE="Register Your Mac"
+    REGISTRATION_TITLE="Register Mac at Organization"
+
+  # Registration status bar text
+    REGISTRATION_STATUS="Waiting on completion of computer registration"
 
   # Registration window submit or finish button text
-    REGISTER_BUTTON="Register"
+    REGISTRATION_BUTTON="Register Your Mac"
 
   # The text and pick list sections below will write the following lines out for
   # end users. Use the variables below to configure what the sentence says
   # Ex: Setting Computer Name to macBook0132
-    REGISTER_BEGIN_STATUS="Setting"
-    REGISTER_MIDDLE_STATUS="to"
+    REGISTRATION_BEGIN_WORD="Setting"
+    REGISTRATION_MIDDLE_WORD="to"
 
   # Registration window can have up to two text fields. Leaving the text display
   # variable empty will hide the input box. Display text is to the side of the
   # input and placeholder text is they grey text inside the input box.
-
-  # Registration window can have up to two dropdown / pick list inputs. Leaving
+  # Registration window can have up to four dropdown / pick list inputs. Leaving
   # the pick display variable empty will hide the dropdown / pick list.
 
-  # Upper Text Field
+  # First Text Field
   #######################################################################################
-    TEXT_UPPER_DISPLAY="Computer Name"
-    TEXT_UPPER_PLACEHOLDER="mac0128371"
+    # Text Field Label
+      REG_TEXT_LABEL_1="Computer Name"
+
+    # Place Holder Text
+      REG_TEXT_LABEL_1_PLACEHOLDER="macBook0123"
+
+    # Optional flag for making the field an optional input for end user
+      REG_TEXT_LABEL_1_OPTIONAL="false" # Set variable to true or false
+
+    # Help Bubble for Input. If title left blank, this will not appear
+      REG_TEXT_LABEL_1_HELP_TITLE="Computer Name Field"
+      REG_TEXT_LABEL_1_HELP_TEXT="This field is sets the name of your new Mac to what is in the Computer Name box. This is important for inventory purposes."
 
     # Logic below was put in this section rather than in core code as folks may
     # want to change what the field does. This is a function that gets called
     # when needed later on. BE VERY CAREFUL IN CHANGING THE FUNCTION!
-      TEXT_UPPER_LOGIC (){
-        TEXT_UPPER_VALUE=$(defaults read "$DEP_NOTIFY_INFO_PLIST" "$TEXT_UPPER_DISPLAY")
-        echo "Status: $REGISTER_BEGIN_STATUS $TEXT_UPPER_DISPLAY $REGISTER_MIDDLE_STATUS $TEXT_UPPER_VALUE" >> "$DEP_NOTIFY_LOG"
-        if [ "$TESTING_MODE" = true ]; then
-           sleep 10
-        else
-          "$JAMF_BINARY" setComputerName -name "$TEXT_UPPER_VALUE"
+      REG_TEXT_LABEL_1_LOGIC (){
+        REG_TEXT_LABEL_1_VALUE=$(defaults read "$DEP_NOTIFY_USER_INPUT_PLIST" "$REG_TEXT_LABEL_1")
+        if [ "$REG_TEXT_LABEL_1_OPTIONAL" = true ] && [ "$REG_TEXT_LABEL_1_VALUE" = "" ]; then
+          echo "Status: $REG_TEXT_LABEL_1 was left empty. Skipping..." >> "$DEP_NOTIFY_LOG"
+          echo "$(date "+%a %h %d %H:%M:%S"): $REG_TEXT_LABEL_1 was set to optional and was left empty. Skipping..." >> "$DEP_NOTIFY_DEBUG"
           sleep 5
+        else
+          echo "Status: $REGISTRATION_BEGIN_WORD $REG_TEXT_LABEL_1 $REGISTRATION_MIDDLE_WORD $REG_TEXT_LABEL_1_VALUE" >> "$DEP_NOTIFY_LOG"
+          if [ "$TESTING_MODE" = true ]; then
+            sleep 10
+          else
+            "$JAMF_BINARY" setComputerName -name "$REG_TEXT_LABEL_1_VALUE"
+            sleep 5
+          fi
         fi
       }
 
-  # Lower Text Field
+  # Second Text Field
   #######################################################################################
-    TEXT_LOWER_DISPLAY="Asset Tag"
-    TEXT_LOWER_PLACEHOLDER="1234567890"
+    # Text Field Label
+      REG_TEXT_LABEL_2="Asset Tag"
+
+    # Place Holder Text
+      REG_TEXT_LABEL_2_PLACEHOLDER="81926392"
+
+    # Optional flag for making the field an optional input for end user
+      REG_TEXT_LABEL_2_OPTIONAL="true" # Set variable to true or false
+
+    # Help Bubble for Input. If title left blank, this will not appear
+      REG_TEXT_LABEL_2_HELP_TITLE="Asset Tag Field"
+      REG_TEXT_LABEL_2_HELP_TEXT="This field is used to give an updated asset tag to our asset management system. If you do not know your asset tag number, please skip this field."
 
     # Logic below was put in this section rather than in core code as folks may
     # want to change what the field does. This is a function that gets called
     # when needed later on. BE VERY CAREFUL IN CHANGING THE FUNCTION!
-      TEXT_LOWER_LOGIC (){
-        TEXT_LOWER_VALUE=$(defaults read "$DEP_NOTIFY_INFO_PLIST" "$TEXT_LOWER_DISPLAY")
-        echo "Status: $REGISTER_BEGIN_STATUS $TEXT_LOWER_DISPLAY $REGISTER_MIDDLE_STATUS $TEXT_LOWER_VALUE" >> "$DEP_NOTIFY_LOG"
-        if [ "$TESTING_MODE" = true ]; then
-           sleep 10
+      REG_TEXT_LABEL_2_LOGIC (){
+        REG_TEXT_LABEL_2_VALUE=$(defaults read "$DEP_NOTIFY_USER_INPUT_PLIST" "$REG_TEXT_LABEL_2")
+        if [ "$REG_TEXT_LABEL_2_OPTIONAL" = true ] && [ "$REG_TEXT_LABEL_2_VALUE" = "" ]; then
+          echo "Status: $REG_TEXT_LABEL_2 was left empty. Skipping..." >> "$DEP_NOTIFY_LOG"
+          echo "$(date "+%a %h %d %H:%M:%S"): $REG_TEXT_LABEL_2 was set to optional and was left empty. Skipping..." >> "$DEP_NOTIFY_DEBUG"
+          sleep 5
         else
-          "$JAMF_BINARY" recon -assetTag "$TEXT_LOWER_VALUE"
+          echo "Status: $REGISTRATION_BEGIN_WORD $REG_TEXT_LABEL_2 $REGISTRATION_MIDDLE_WORD $REG_TEXT_LABEL_2_VALUE" >> "$DEP_NOTIFY_LOG"
+          if [ "$TESTING_MODE" = true ]; then
+             sleep 10
+          else
+            "$JAMF_BINARY" recon -assetTag "$REG_TEXT_LABEL_2_VALUE"
+          fi
         fi
       }
 
-  # Upper Dropdown
+  # Popup 1
   #######################################################################################
-    PICK_UPPER_DISPLAY="Building"
-    PICK_UPPER_OPTIONS=(
-      "Amsterdam"
-      "Eau Claire"
-      "Minneapolis"
-    )
+    # Label for the popup
+      REG_POPUP_LABEL_1="Building"
+
+    # Array of options for the user to select
+      REG_POPUP_LABEL_1_OPTIONS=(
+        "Amsterdam"
+        "Eau Claire"
+        "Minneapolis"
+      )
+
+    # Help Bubble for Input. If title left blank, this will not appear
+      REG_POPUP_LABEL_1_HELP_TITLE="Building Dropdown Field"
+      REG_POPUP_LABEL_1_HELP_TEXT="Please choose the appropriate building for where you normally work. This is important for inventory purposes."
 
     # Logic below was put in this section rather than in core code as folks may
     # want to change what the field does. This is a function that gets called
     # when needed later on. BE VERY CAREFUL IN CHANGING THE FUNCTION!
-      PICK_UPPER_LOGIC (){
-        PICK_UPPER_VALUE=$(defaults read "$DEP_NOTIFY_INFO_PLIST" "$PICK_UPPER_DISPLAY")
-        echo "Status: $REGISTER_BEGIN_STATUS $PICK_UPPER_DISPLAY $REGISTER_MIDDLE_STATUS $PICK_UPPER_VALUE" >> "$DEP_NOTIFY_LOG"
+      REG_POPUP_LABEL_1_LOGIC (){
+        REG_POPUP_LABEL_1_VALUE=$(defaults read "$DEP_NOTIFY_USER_INPUT_PLIST" "$REG_POPUP_LABEL_1")
+        echo "Status: $REGISTRATION_BEGIN_WORD $REG_POPUP_LABEL_1 $REGISTRATION_MIDDLE_WORD $REG_POPUP_LABEL_1_VALUE" >> "$DEP_NOTIFY_LOG"
         if [ "$TESTING_MODE" = true ]; then
            sleep 10
         else
-          "$JAMF_BINARY" recon -building "$PICK_UPPER_VALUE"
+          "$JAMF_BINARY" recon -building "$REG_POPUP_LABEL_1_VALUE"
         fi
       }
 
-  # Lower Dropdown
+  # Popup 2
   #######################################################################################
-    PICK_LOWER_DISPLAY="Department"
-    PICK_LOWER_OPTIONS=(
-      "Customer Onboarding"
-      "Professional Services"
-      "Sales Engineering"
-    )
+    # Label for the popup
+      REG_POPUP_LABEL_2="Department"
+
+    # Array of options for the user to select
+      REG_POPUP_LABEL_2_OPTIONS=(
+        "Customer Onboarding"
+        "Professional Services"
+        "Sales Engineering"
+      )
+
+    # Help Bubble for Input. If title left blank, this will not appear
+      REG_POPUP_LABEL_2_HELP_TITLE="Department Dropdown Field"
+      REG_POPUP_LABEL_2_HELP_TEXT="Please choose the appropriate department for where you normally work. This is important for inventory purposes."
 
     # Logic below was put in this section rather than in core code as folks may
     # want to change what the field does. This is a function that gets called
     # when needed later on. BE VERY CAREFUL IN CHANGING THE FUNCTION!
-      PICK_LOWER_LOGIC (){
-        PICK_LOWER_VALUE=$(defaults read "$DEP_NOTIFY_INFO_PLIST" "$PICK_LOWER_DISPLAY")
-        echo "Status: $REGISTER_BEGIN_STATUS $PICK_LOWER_DISPLAY $REGISTER_MIDDLE_STATUS $PICK_LOWER_VALUE" >> "$DEP_NOTIFY_LOG"
+      REG_POPUP_LABEL_2_LOGIC (){
+        REG_POPUP_LABEL_2_VALUE=$(defaults read "$DEP_NOTIFY_USER_INPUT_PLIST" "$REG_POPUP_LABEL_2")
+        echo "Status: $REGISTRATION_BEGIN_WORD $REG_POPUP_LABEL_2 $REGISTRATION_MIDDLE_WORD $REG_POPUP_LABEL_2_VALUE" >> "$DEP_NOTIFY_LOG"
         if [ "$TESTING_MODE" = true ]; then
            sleep 10
         else
-          "$JAMF_BINARY" recon -department "$PICK_LOWER_VALUE"
+          "$JAMF_BINARY" recon -department "$REG_POPUP_LABEL_2_VALUE"
+        fi
+      }
+
+  # Popup 3 - Code is here but currently unused
+  #######################################################################################
+    # Label for the popup
+      REG_POPUP_LABEL_3=""
+
+    # Array of options for the user to select
+      REG_POPUP_LABEL_3_OPTIONS=(
+        "Option 1"
+        "Option 2"
+        "Option 3"
+      )
+
+    # Help Bubble for Input. If title left blank, this will not appear
+      REG_POPUP_LABEL_3_HELP_TITLE="Dropdown 3 Field"
+      REG_POPUP_LABEL_3_HELP_TEXT="This dropdown is currently not in use. All code is here ready for you to use. It can also be hidden by removing the contents of the REG_POPUP_LABEL_3 variable."
+
+    # Logic below was put in this section rather than in core code as folks may
+    # want to change what the field does. This is a function that gets called
+    # when needed later on. BE VERY CAREFUL IN CHANGING THE FUNCTION!
+      REG_POPUP_LABEL_3_LOGIC (){
+        REG_POPUP_LABEL_3_VALUE=$(defaults read "$DEP_NOTIFY_USER_INPUT_PLIST" "$REG_POPUP_LABEL_3")
+        echo "Status: $REGISTRATION_BEGIN_WORD $REG_POPUP_LABEL_3 $REGISTRATION_MIDDLE_WORD $REG_POPUP_LABEL_3_VALUE" >> "$DEP_NOTIFY_LOG"
+        if [ "$TESTING_MODE" = true ]; then
+          sleep 10
+        else
+          sleep 10
+        fi
+      }
+
+  # Popup 4 - Code is here but currently unused
+  #######################################################################################
+    # Label for the popup
+      REG_POPUP_LABEL_4=""
+
+    # Array of options for the user to select
+      REG_POPUP_LABEL_4_OPTIONS=(
+        "Option 1"
+        "Option 2"
+        "Option 3"
+      )
+
+    # Help Bubble for Input. If title left blank, this will not appear
+      REG_POPUP_LABEL_4_HELP_TITLE="Dropdown 4 Field"
+      REG_POPUP_LABEL_4_HELP_TEXT="This dropdown is currently not in use. All code is here ready for you to use. It can also be hidden by removing the contents of the REG_POPUP_LABEL_4 variable."
+
+    # Logic below was put in this section rather than in core code as folks may
+    # want to change what the field does. This is a function that gets called
+    # when needed later on. BE VERY CAREFUL IN CHANGING THE FUNCTION!
+      REG_POPUP_LABEL_4_LOGIC (){
+        REG_POPUP_LABEL_4_VALUE=$(defaults read "$DEP_NOTIFY_USER_INPUT_PLIST" "$REG_POPUP_LABEL_4")
+        echo "Status: $REGISTRATION_BEGIN_WORD $REG_POPUP_LABEL_4 $REGISTRATION_MIDDLE_WORD $REG_POPUP_LABEL_4_VALUE" >> "$DEP_NOTIFY_LOG"
+        if [ "$TESTING_MODE" = true ]; then
+          sleep 10
+        else
+          sleep 10
         fi
       }
 
 #########################################################################################
+#########################################################################################
 # Core Script Logic - Don't Change Without Major Testing
+#########################################################################################
 #########################################################################################
 
 # Variables for File Paths
@@ -304,7 +440,7 @@
   # EULA Mode
     if [ "$9" != "" ]; then EULA_ENABLED="$9"; fi
   # Registration Mode
-    if [ "${10}" != "" ]; then REGISTER_ENABLED="${10}"; fi
+    if [ "${10}" != "" ]; then REGISTRATION_ENABLED="${10}"; fi
 
 # Standard Testing Mode Enhancements
   if [ "$TESTING_MODE" = true ]; then
@@ -341,8 +477,8 @@
     echo "$(date "+%a %h %d %H:%M:%S"): EULA configuration not set properly. Currently set to $EULA_ENABLED. Please update to true or false." >> "$DEP_NOTIFY_DEBUG"
     exit 1
   fi
-  if [ "$REGISTER_ENABLED" != true ] && [ "$REGISTER_ENABLED" != false ]; then
-    echo "$(date "+%a %h %d %H:%M:%S"): Registration configuration not set properly. Currently set to $REGISTER_ENABLED. Please update to true or false." >> "$DEP_NOTIFY_DEBUG"
+  if [ "$REGISTRATION_ENABLED" != true ] && [ "$REGISTRATION_ENABLED" != false ]; then
+    echo "$(date "+%a %h %d %H:%M:%S"): Registration configuration not set properly. Currently set to $REGISTRATION_ENABLED. Please update to true or false." >> "$DEP_NOTIFY_DEBUG"
     exit 1
   fi
 
@@ -408,23 +544,28 @@
 # Setting custom main text if specified
   if [ "$MAIN_TEXT" != "" ]; then echo "Command: MainText: $MAIN_TEXT" >> "$DEP_NOTIFY_LOG"; fi
 
-# Adding help url and button if specified
-  if [ "$SUPPORT_URL" != "" ]; then echo "Command: Help: $SUPPORT_URL" >> "$DEP_NOTIFY_LOG"; fi
-
-# Plist Location configuration
+# General Plist Configuration
   # Calling function to set the INFO_PLIST_PATH
     INFO_PLIST_WRAPPER
 
-  # The plist information below is used by EULA and Registration windows
+  # The plist information below
     DEP_NOTIFY_CONFIG_PLIST="/Users/$CURRENT_USER/Library/Preferences/menu.nomad.DEPNotify.plist"
-    DEP_NOTIFY_INFO_PLIST="$DEP_NOTIFY_INFO_PLIST_PATH/DEPNotify.plist"
 
   # If testing mode is on, this will remove some old configuration files
     if [ "$TESTING_MODE" = true ] && [ -f "$DEP_NOTIFY_CONFIG_PLIST" ]; then rm "$DEP_NOTIFY_CONFIG_PLIST"; fi
-    if [ "$TESTING_MODE" = true ] && [ -f "$DEP_NOTIFY_INFO_PLIST" ]; then rm "$DEP_NOTIFY_INFO_PLIST"; fi
+    if [ "$TESTING_MODE" = true ] && [ -f "$DEP_NOTIFY_USER_INPUT_PLIST" ]; then rm "$DEP_NOTIFY_USER_INPUT_PLIST"; fi
 
   # Setting default path to the plist which stores all the user completed info
-    defaults write "$DEP_NOTIFY_CONFIG_PLIST" PathToPlistFile "$DEP_NOTIFY_INFO_PLIST_PATH"
+    defaults write "$DEP_NOTIFY_CONFIG_PLIST" pathToPlistFile "$DEP_NOTIFY_USER_INPUT_PLIST"
+
+  # Setting status text alignment
+    defaults write "$DEP_NOTIFY_CONFIG_PLIST" statusTextAlignment "$STATUS_TEXT_ALIGN"
+
+  # Setting help button
+    if [ "$HELP_BUBBLE_TITLE" != "" ]; then
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" helpBubble -array-add "$HELP_BUBBLE_TITLE"
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" helpBubble -array-add "$HELP_BUBBLE_BODY"
+    fi
 
 # EULA Configuration
   if [ "$EULA_ENABLED" =  true ]; then
@@ -433,52 +574,113 @@
     # If testing mode is on, this will remove EULA specific configuration files
       if [ "$TESTING_MODE" = true ] && [ -f "$DEP_NOTIFY_EULA_DONE" ]; then rm "$DEP_NOTIFY_EULA_DONE"; fi
 
-    # Writing the location of the EULA file and changing ownership of the file
+    # Writing title, subtitle, and EULA txt location to plist
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" EULAMainTitle "$EULA_MAIN_TITLE"
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" EULASubTitle "$EULA_SUBTITLE"
       defaults write "$DEP_NOTIFY_CONFIG_PLIST" pathToEULA "$EULA_FILE_PATH"
-      chown "$CURRENT_USER" "$EULA_FILE_PATH"
-      chown "$CURRENT_USER" "$DEP_NOTIFY_CONFIG_PLIST"
+
+    # Setting ownership of files
+      chown "$CURRENT_USER:staff" "$EULA_FILE_PATH"
+      chmod 444 "$EULA_FILE_PATH"
+      chown "$CURRENT_USER:staff" "$DEP_NOTIFY_CONFIG_PLIST"
+      chmod 600 "$DEP_NOTIFY_CONFIG_PLIST"
   fi
 
 # Registration Plist Configuration
-  if [ "$REGISTER_ENABLED" = true ]; then
+  if [ "$REGISTRATION_ENABLED" = true ]; then
     DEP_NOTIFY_REGISTER_DONE="/var/tmp/com.depnotify.registration.done"
 
     # If testing mode is on, this will remove registration specific configuration files
       if [ "$TESTING_MODE" = true ] && [ -f "$DEP_NOTIFY_REGISTER_DONE" ]; then rm "$DEP_NOTIFY_REGISTER_DONE"; fi
 
     # Main Window Text Configuration
-      defaults write "$DEP_NOTIFY_CONFIG_PLIST" RegisterMainTitle "$REGISTER_TITLE"
-      defaults write "$DEP_NOTIFY_CONFIG_PLIST" RegisterButtonLabel "$REGISTER_BUTTON"
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" registrationMainTitle "$REGISTRATION_TITLE"
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" registrationButtonLabel "$REGISTRATION_BUTTON"
+      defaults write "$DEP_NOTIFY_CONFIG_PLIST" registrationPicturePath "$BANNER_IMAGE_PATH"
 
-    # Upper Text Box Configuration
-      if [ "$TEXT_UPPER_DISPLAY" != "" ]; then
-        defaults write "$DEP_NOTIFY_CONFIG_PLIST" UITextFieldUpperLabel "$TEXT_UPPER_DISPLAY"
-        defaults write "$DEP_NOTIFY_CONFIG_PLIST" UITextFieldUpperPlaceholder "$TEXT_UPPER_PLACEHOLDER"
+    # First Text Box Configuration
+      if [ "$REG_TEXT_LABEL_1" != "" ]; then
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField1Label "$REG_TEXT_LABEL_1"
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField1Placeholder "$REG_TEXT_LABEL_1_PLACEHOLDER"
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField1IsOptional "$REG_TEXT_LABEL_1_OPTIONAL"
+        # Code for showing the help box if configured
+          if [ "$REG_TEXT_LABEL_1_HELP_TITLE" != "" ]; then
+              defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField1Bubble -array-add "$REG_TEXT_LABEL_1_HELP_TITLE"
+              defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField1Bubble -array-add "$REG_TEXT_LABEL_1_HELP_TEXT"
+          fi
       fi
 
-    # Lower Text Box Configuration
-      if [ "$TEXT_LOWER_DISPLAY" != "" ]; then
-        defaults write "$DEP_NOTIFY_CONFIG_PLIST" UITextFieldLowerLabel "$TEXT_LOWER_DISPLAY"
-        defaults write "$DEP_NOTIFY_CONFIG_PLIST" UITextFieldLowerPlaceholder "$TEXT_LOWER_PLACEHOLDER"
+    # Second Text Box Configuration
+      if [ "$REG_TEXT_LABEL_2" != "" ]; then
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField2Label "$REG_TEXT_LABEL_2"
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField2Placeholder "$REG_TEXT_LABEL_2_PLACEHOLDER"
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField2IsOptional "$REG_TEXT_LABEL_2_OPTIONAL"
+        # Code for showing the help box if configured
+          if [ "$REG_TEXT_LABEL_2_HELP_TITLE" != "" ]; then
+              defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField2Bubble -array-add "$REG_TEXT_LABEL_2_HELP_TITLE"
+              defaults write "$DEP_NOTIFY_CONFIG_PLIST" textField2Bubble -array-add "$REG_TEXT_LABEL_2_HELP_TEXT"
+          fi
       fi
 
-    # Upper Pick List / Dropdown Configuration
-      if [ "$PICK_UPPER_DISPLAY" != "" ]; then
-        defaults write "$DEP_NOTIFY_CONFIG_PLIST" UIPopUpMenuUpperLabel "$PICK_UPPER_DISPLAY"
-        for PICK_UPPER_OPTION in "${PICK_UPPER_OPTIONS[@]}"; do
-           defaults write "$DEP_NOTIFY_CONFIG_PLIST" UIPopUpMenuUpper -array-add "$PICK_UPPER_OPTION"
-        done
+    # Popup 1
+      if [ "$REG_POPUP_LABEL_1" != "" ]; then
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton1Label "$REG_POPUP_LABEL_1"
+        # Code for showing the help box if configured
+          if [ "$REG_POPUP_LABEL_1_HELP_TITLE" != "" ]; then
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu1Bubble -array-add "$REG_POPUP_LABEL_1_HELP_TITLE"
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu1Bubble -array-add "$REG_POPUP_LABEL_1_HELP_TEXT"
+          fi
+        # Code for adding the items from the array above into the plist
+          for REG_POPUP_LABEL_1_OPTION in "${REG_POPUP_LABEL_1_OPTIONS[@]}"; do
+             defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton1Content -array-add "$REG_POPUP_LABEL_1_OPTION"
+          done
       fi
 
-    # Lower Pick List / Dropdown Configuration
-      if [ "$PICK_LOWER_DISPLAY" != "" ]; then
-        defaults write "$DEP_NOTIFY_CONFIG_PLIST" UIPopUpMenuLowerLabel "$PICK_LOWER_DISPLAY"
-        for PICK_LOWER_OPTION in "${PICK_LOWER_OPTIONS[@]}"; do
-           defaults write "$DEP_NOTIFY_CONFIG_PLIST" UIPopUpMenuLower -array-add "$PICK_LOWER_OPTION"
-        done
+    # Popup 2
+      if [ "$REG_POPUP_LABEL_2" != "" ]; then
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton2Label "$REG_POPUP_LABEL_2"
+        # Code for showing the help box if configured
+          if [ "$REG_POPUP_LABEL_2_HELP_TITLE" != "" ]; then
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu2Bubble -array-add "$REG_POPUP_LABEL_2_HELP_TITLE"
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu2Bubble -array-add "$REG_POPUP_LABEL_2_HELP_TEXT"
+          fi
+        # Code for adding the items from the array above into the plist
+          for REG_POPUP_LABEL_2_OPTION in "${REG_POPUP_LABEL_2_OPTIONS[@]}"; do
+             defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton2Content -array-add "$REG_POPUP_LABEL_2_OPTION"
+          done
       fi
+
+    # Popup 3
+      if [ "$REG_POPUP_LABEL_3" != "" ]; then
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton3Label "$REG_POPUP_LABEL_3"
+        # Code for showing the help box if configured
+          if [ "$REG_POPUP_LABEL_3_HELP_TITLE" != "" ]; then
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu3Bubble -array-add "$REG_POPUP_LABEL_3_HELP_TITLE"
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu3Bubble -array-add "$REG_POPUP_LABEL_3_HELP_TEXT"
+          fi
+        # Code for adding the items from the array above into the plist
+          for REG_POPUP_LABEL_3_OPTION in "${REG_POPUP_LABEL_3_OPTIONS[@]}"; do
+             defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton3Content -array-add "$REG_POPUP_LABEL_3_OPTION"
+          done
+      fi
+
+    # Popup 4
+      if [ "$REG_POPUP_LABEL_4" != "" ]; then
+        defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton4Label "$REG_POPUP_LABEL_4"
+        # Code for showing the help box if configured
+          if [ "$REG_POPUP_LABEL_4_HELP_TITLE" != "" ]; then
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu4Bubble -array-add "$REG_POPUP_LABEL_4_HELP_TITLE"
+            defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupMenu4Bubble -array-add "$REG_POPUP_LABEL_4_HELP_TEXT"
+          fi
+        # Code for adding the items from the array above into the plist
+          for REG_POPUP_LABEL_4_OPTION in "${REG_POPUP_LABEL_4_OPTIONS[@]}"; do
+             defaults write "$DEP_NOTIFY_CONFIG_PLIST" popupButton4Content -array-add "$REG_POPUP_LABEL_4_OPTION"
+          done
+      fi
+
     # Changing Ownership of the plist file
-      chown "$CURRENT_USER" "$DEP_NOTIFY_CONFIG_PLIST"
+      chown "$CURRENT_USER":staff "$DEP_NOTIFY_CONFIG_PLIST"
+      chmod 600 "$DEP_NOTIFY_CONFIG_PLIST"
   fi
 
 # Opening the app after initial configuration
@@ -516,11 +718,13 @@
   # more based on EULA, register, or other options.
     ADDITIONAL_OPTIONS_COUNTER=1
     if [ "$EULA_ENABLED" = true ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
-    if [ "$REGISTER_ENABLED" = true ]; then ((ADDITIONAL_OPTIONS_COUNTER++))
-      if [ "$TEXT_UPPER_DISPLAY" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
-      if [ "$TEXT_LOWER_DISPLAY" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
-      if [ "$PICK_UPPER_DISPLAY" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
-      if [ "$PICK_LOWER_DISPLAY" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
+    if [ "$REGISTRATION_ENABLED" = true ]; then ((ADDITIONAL_OPTIONS_COUNTER++))
+      if [ "$REG_TEXT_LABEL_1" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
+      if [ "$REG_TEXT_LABEL_2" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
+      if [ "$REG_POPUP_LABEL_1" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
+      if [ "$REG_POPUP_LABEL_2" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
+      if [ "$REG_POPUP_LABEL_3" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
+      if [ "$REG_POPUP_LABEL_4" != "" ]; then ((ADDITIONAL_OPTIONS_COUNTER++)); fi
     fi
 
   # Checking policy array and adding the count from the additional options above.
@@ -529,25 +733,29 @@
 
 # EULA Window Display Logic
   if [ "$EULA_ENABLED" = true ]; then
-    echo "Status: Waiting on EULA Acceptance" >> "$DEP_NOTIFY_LOG"
-    echo "Command: ContinueButtonEULA: EULA" >> "$DEP_NOTIFY_LOG"
+    echo "Status: $EULA_STATUS" >> "$DEP_NOTIFY_LOG"
+    echo "Command: ContinueButtonEULA: $EULA_BUTTON" >> "$DEP_NOTIFY_LOG"
     while [ ! -f "$DEP_NOTIFY_EULA_DONE" ]; do
+      echo "$(date "+%a %h %d %H:%M:%S"): Waiting for user to accept EULA." >> "$DEP_NOTIFY_DEBUG"
       sleep 1
     done
   fi
 
 # Registration Window Display Logic
-  if [ "$REGISTER_ENABLED" = true ]; then
-    echo "Status: $REGISTER_TITLE" >> "$DEP_NOTIFY_LOG"
-    echo "Command: ContinueButtonRegister: Register" >> "$DEP_NOTIFY_LOG"
+  if [ "$REGISTRATION_ENABLED" = true ]; then
+    echo "Status: $REGISTRATION_STATUS" >> "$DEP_NOTIFY_LOG"
+    echo "Command: ContinueButtonRegister: $REGISTRATION_BUTTON" >> "$DEP_NOTIFY_LOG"
     while [ ! -f "$DEP_NOTIFY_REGISTER_DONE" ]; do
+      echo "$(date "+%a %h %d %H:%M:%S"): Waiting for user to complete registration." >> "$DEP_NOTIFY_DEBUG"
       sleep 1
     done
     # Running Logic For Each Registration Box
-      if [ "$TEXT_UPPER_DISPLAY" != "" ]; then TEXT_UPPER_LOGIC; fi
-      if [ "$TEXT_LOWER_DISPLAY" != "" ]; then TEXT_LOWER_LOGIC; fi
-      if [ "$PICK_UPPER_DISPLAY" != "" ]; then PICK_UPPER_LOGIC; fi
-      if [ "$PICK_LOWER_DISPLAY" != "" ]; then PICK_LOWER_LOGIC; fi
+      if [ "$REG_TEXT_LABEL_1" != "" ]; then REG_TEXT_LABEL_1_LOGIC; fi
+      if [ "$REG_TEXT_LABEL_2" != "" ]; then REG_TEXT_LABEL_2_LOGIC; fi
+      if [ "$REG_POPUP_LABEL_1" != "" ]; then REG_POPUP_LABEL_1_LOGIC; fi
+      if [ "$REG_POPUP_LABEL_2" != "" ]; then REG_POPUP_LABEL_2_LOGIC; fi
+      if [ "$REG_POPUP_LABEL_3" != "" ]; then REG_POPUP_LABEL_3_LOGIC; fi
+      if [ "$REG_POPUP_LABEL_4" != "" ]; then REG_POPUP_LABEL_4_LOGIC; fi
   fi
 
 # Loop to run policies