+++ CHANGELOG.md
@@ -1,5 +1,64 @@
 # Change Log
 
+## [2.0.0] - 2018/12/10
+
+Moved to new version release as this release does not have backwards compatibility for older versions of DEPNotify. Please review the new [RELEASES.md](RELEASES.md) page for more information.
+
+* Added [RELEASES.md](RELEASES.md) page to track versions of script app and macOS that have been tested together
+* Made small modifications to the [README.md](README.md) page to alert admins to new page
+* Separated configurable variables into small chunks by use. Required configurations are closer to the top with more niche options closer to the bottom of the configuration area
+* Removed `SUPPORT_URL` as newer version does not support
+* Added `STATUS_TEXT_ALIGN`
+* Added `HELP_BUBBLE_TITLE`
+* Added `HELP_BUBBLE_BODY`
+* `DEP_NOTIFY_INFO_PLIST` changed to `DEP_NOTIFY_USER_INPUT_PLIST` as there is a new plist that stores all user input separate from configuration
+* EULA Enhancements
+  * Added `EULA_MAIN_TITLE`
+  * Added `EULA_STATUS`
+  * Added `EULA_BUTTON`
+  * Added `EULA_SUBTITLE`
+* Registration Enhancements
+  * `REGISTER_ENABLED` changed to `REGISTRATION_ENABLED` to keep registration options similar
+  * Added `REGISTRATION_STATUS`
+  * `REGISTER_TITLE` changed to `REGISTRATION_TITLE` to keep registration options similar
+  * `REGISTER_BEGIN_STATUS` changed to `REGISTRATION_BEGIN_WORD`
+  * `REGISTER_MIDDLE_STATUS` changed to `REGISTRATION_MIDDLE_WORD`
+  * First text box
+    * `TEXT_UPPER_DISPLAY` changed to `REG_TEXT_LABEL_1`
+    * `TEXT_UPPER_PLACEHOLDER` changed to `REG_TEXT_LABEL_1_PLACEHOLDER`
+    * Added `REG_TEXT_LABEL_1_OPTIONAL`
+    * Added code to allow for optional items to be skipped if empty
+    * Added `REG_TEXT_LABEL_1_HELP_TITLE`
+    * Added `REG_TEXT_LABEL_1_HELP_TEXT`
+    * `TEXT_UPPER_LOGIC` changed to `REG_TEXT_LABEL_1_LOGIC`
+    * `TEXT_UPPER_VALUE` changed to `REG_TEXT_LABEL_1_VALUE`
+  * Second text box
+    * `TEXT_LOWER_DISPLAY` changed to `REG_TEXT_LABEL_2`
+    * `TEXT_LOWER_PLACEHOLDER` changed to `REG_TEXT_LABEL_2_PLACEHOLDER`
+    * Added `REG_TEXT_LABEL_2_OPTIONAL`
+    * Added code to allow for optional items to be skipped if empty
+    * Added `REG_TEXT_LABEL_2_HELP_TITLE`
+    * Added `REG_TEXT_LABEL_2_HELP_TEXT`
+    * `TEXT_LOWER_LOGIC` changed to `REG_TEXT_LABEL_2_LOGIC`
+    * `TEXT_LOWER_VALUE` changed to `REG_TEXT_LABEL_2_VALUE`
+  * Popup 1
+    * `PICK_UPPER_DISPLAY` changed to `REG_POPUP_LABEL_1`
+    * `PICK_UPPER_OPTIONS` changed to `REG_POPUP_LABEL_1_OPTIONS`
+    * Added `REG_POPUP_LABEL_1_HELP_TITLE`
+    * Added `REG_POPUP_LABEL_1_HELP_TEXT`
+    * `PICK_UPPER_LOGIC` changed to `REG_POPUP_LABEL_1_LOGIC`
+    * `PICK_UPPER_VALUE` changed to `REG_POPUP_LABEL_1_VALUE`
+  * Popup 2
+    * `PICK_LOWER_DISPLAY` changed to `REG_POPUP_LABEL_2`
+    * `PICK_LOWER_OPTIONS` changed to `REG_POPUP_LABEL_2_OPTIONS`
+    * Added `REG_POPUP_LABEL_2_HELP_TITLE`
+    * Added `REG_POPUP_LABEL_2_HELP_TEXT`
+    * `PICK_LOWER_LOGIC` changed to `REG_POPUP_LABEL_2_LOGIC`
+    * `PICK_LOWER_VALUE` changed to `REG_POPUP_LABEL_2_VALUE`
+  * Added full code to support popup 3
+  * Added full code to support popup 4
+  * Modified code for ownership and permissions on plists within the user's library folder
+
 ## [1.2.0] - 2018/11/01
 
 Major enhancements and additions across the board by Kyle Bareis